fix mux
